Create traces feed resource
